requirements.txt
streamlit
opencv-python
secure-smtplib
opencv-python-headless
pillow
requests
geopy
SpeechRecognition
pyAudio